backend/app/schemas/user.py
# Standard library
from datetime import datetime

# Third-party
from pydantic import BaseModel, EmailStr


class UserResponse(BaseModel):
    id: int
    email: EmailStr
    role: str
    is_active: bool
    created_at: datetime
