backend/pyproject.toml
[tool.poetry]
name = "events"
version = "0.1.0"
description = "FastAPI for events"
authors = ["Jeiikot <jeiikot21@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlmodel = "^0.0.22"
SQLAlchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.3.4"
python-jose = "^3.3.0"
passlib = {version="^1.7.4", extras=["bcrypt"]}
alembic = "^1.13.2"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
